Use SCOPE_IDENTITY() to get inserted id

+++ src/main/java/com/epam/fitnesstracker/domain/User.java
@@ -2,7 +2,7 @@ package com.epam.fitnesstracker.domain;
 
 public class User {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -12,17 +12,17 @@ public class User {
 
     }
 
-    public User(Integer id, String name, Integer age) {
+    public User(Long id, String name, Integer age) {
         this.id = id;
         this.name = name;
         this.age = age;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
